Orthography WebAssembly: cleanup

+++ Orthography2/Models/Word.cs
@@ -4,7 +4,7 @@
 	{
 		public int ID { get; set; }
 		public string Value { get; set; }
-		public int PrepositionsMask { get; set; }
+		public decimal PrepositionsMask { get; set; }
 		public string Translation { get; set; }
 	}
 }
